+++ source/win-preferences/schema/display.ts
@@ -118,6 +118,13 @@ export default function (): any {
       ],
       [
         {
+          type: 'checkbox',
+          label: trans('dialog.preferences.display.hide_toolbar_distraction_free'),
+          model: 'display.hideToolbarInDistractionFree'
+        }
+      ],
+      [
+        {
           type: 'fieldset-label', // TODO: Create this type
           text: trans('dialog.preferences.display.image_size_info')
         },